requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
qdrant-client==1.9.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
SpeechRecognition==3.10.0
spacy==3.7.2
nltk==3.8.1
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
pytesseract==0.3.10
scikit-learn==1.3.2
numpy==1.24.3
